+++ src/agent/cmd.ts
@@ -1,4 +1,4 @@
-import {pick, uniqBy} from "ramda"
+import {last, pick, uniqBy} from "ramda"
 import {get} from "svelte/store"
 import {doPipe} from "hurdak/lib/hurdak"
 import {parseContent, Tags, roomAttrs, displayPerson, findRoot, findReply} from "src/util/nostr"
@@ -84,6 +84,8 @@ const requestZap = (relays, content, pubkey, eventId, amount, lnurl) => {
 
 const deleteEvent = ids => new PublishableEvent(5, {tags: ids.map(id => ["e", id])})
 
+const createLabel = payload => new PublishableEvent(1985, payload)
+
 // Utils
 
 const tagsFromContent = (content, tags) => {
@@ -120,6 +122,7 @@ const getReplyTags = (n, inherit = false) => {
   const extra = inherit
     ? Tags.from(n)
         .type("e")
+        .reject(t => last(t) === "mention")
         .all()
         .map(t => t.slice(0, 3))
     : []
@@ -158,7 +161,7 @@ class PublishableEvent {
   }
   async publish(relays, onProgress = null, verb = "EVENT") {
     const event = await this.getSignedEvent()
-    // console.log(event); return
+    // return console.log(event)
     const promise = pool.publish({relays, event, onProgress, verb})
 
     // Copy the event since loki mutates it to add metadata
@@ -186,5 +189,6 @@ export default {
   createReply,
   requestZap,
   deleteEvent,
+  createLabel,
   PublishableEvent,
 }