+++ src/app/ModalRoutes.svelte
@@ -19,8 +19,9 @@
   import ListList from "src/app/views/ListList.svelte"
   import ListSelect from "src/app/views/ListSelect.svelte"
   import ListEdit from "src/app/views/ListEdit.svelte"
-  import RelayAdd from "src/app/views/RelayAdd.svelte"
+  import RelayBrowse from "src/app/views/RelayBrowse.svelte"
   import RelayDetail from "src/app/views/RelayDetail.svelte"
+  import RelayReview from "src/app/views/RelayReview.svelte"
 
   export let m
 </script>
@@ -37,10 +38,12 @@
   <NoteZap note={m.note} />
 {:else if m.type === "note/share"}
   <NoteShare note={m.note} />
-{:else if m.type === "relay/add"}
-  <RelayAdd url={m.url} />
+{:else if m.type === "relay/browse"}
+  <RelayBrowse />
 {:else if m.type === "relay/detail"}
   <RelayDetail url={m.url} />
+{:else if m.type === "relay/review"}
+  <RelayReview url={m.url} />
 {:else if m.type === "onboarding"}
   <Onboarding stage={m.stage} />
 {:else if m.type === "room/edit"}