+++ src/app/SideNav.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import cx from "classnames"
   import {displayPerson} from "src/util/nostr"
-  import {installPrompt} from "src/partials/state"
+  import {theme, installPrompt} from "src/partials/state"
   import user from "src/agent/user"
   import pool from "src/agent/pool"
   import {routes, menuIsOpen} from "src/app/state"
@@ -10,6 +10,8 @@
   import PersonCircle from "src/app/shared/PersonCircle.svelte"
 
   const {profile, canPublish} = user
+
+  const toggleTheme = () => theme.update(t => (t === "dark" ? "light" : "dark"))
 
   const install = () => {
     $installPrompt.prompt()
@@ -70,11 +72,7 @@
       {/if}
     </a>
   </li>
-  <li
-    class={cx("relative", {
-      "cursor-pointer": $canPublish,
-      "pointer-events-none opacity-75": !$canPublish,
-    })}>
+  <li class="relative">
     <a class="block px-4 py-2 transition-all hover:bg-accent hover:text-white" href="/chat">
       <i class="fa fa-comment mr-2" /> Chat
       {#if $newChatMessages}
@@ -109,6 +107,13 @@
         <i class="fa fa-gear mr-2" /> Settings
       </a>
     </li>
+  {/if}
+  <li
+    class="block cursor-pointer px-4 py-2 transition-all hover:bg-accent hover:text-white"
+    on:click={toggleTheme}>
+    <i class="fa fa-lightbulb mr-2" /> Theme
+  </li>
+  {#if $profile.pubkey}
     <li class="cursor-pointer">
       <a class="block px-4 py-2 transition-all hover:bg-accent hover:text-white" href="/logout">
         <i class="fa fa-right-from-bracket mr-2" /> Logout