+++ src/app/shared/NoteContent.svelte
@@ -1,14 +1,16 @@
 <script lang="ts">
   import {objOf, reverse} from "ramda"
   import {fly} from "svelte/transition"
-  import {splice} from "hurdak/lib/hurdak"
+  import {splice, switcher, switcherFn} from "hurdak/lib/hurdak"
   import {warn} from "src/util/logger"
-  import {displayPerson, parseContent, Tags} from "src/util/nostr"
+  import {displayPerson, parseContent, getLabelQuality, displayRelay, Tags} from "src/util/nostr"
   import {modal} from "src/partials/state"
   import MediaSet from "src/partials/MediaSet.svelte"
+  import QRCode from "src/partials/QRCode.svelte"
   import Card from "src/partials/Card.svelte"
   import Spinner from "src/partials/Spinner.svelte"
   import Anchor from "src/partials/Anchor.svelte"
+  import Rating from "src/partials/Rating.svelte"
   import PersonCircle from "src/app/shared/PersonCircle.svelte"
   import {sampleRelays} from "src/agent/relays"
   import user from "src/agent/user"
@@ -25,17 +27,27 @@
   const shouldTruncate = !showEntire && note.content.length > maxLength
 
   let content = parseContent(note)
+  let rating = note.kind === 1985 ? getLabelQuality("review/relay", note) : null
+
+  console.log(content)
 
   const links = []
+  const invoices = []
   const ranges = []
 
   // Find links and preceding whitespace
   for (let i = 0; i < content.length; i++) {
     const {type, value} = content[i]
 
-    if (type === "link" && !value.startsWith("ws")) {
+    if (type === "link") {
       links.push(value)
+    }
 
+    if (type === "lnurl") {
+      invoices.push(value)
+    }
+
+    if (["link", "lnurl"].includes(type) && !value.startsWith("ws")) {
       const prev = content[i - 1]
       const next = content[i + 1]
 
@@ -118,8 +130,35 @@
 
 <div class="flex flex-col gap-2 overflow-hidden text-ellipsis">
   <p>
+    {#if rating}
+      {@const [type, value] = Tags.from(note)
+        .reject(t => ["l", "L"].includes(t[0]))
+        .first()}
+      {@const action = switcher(type, {
+        r: () => modal.push({type: "relay/detail", url: value}),
+        p: () => modal.push({type: "person/feed", pubkey: value}),
+        e: () => modal.push({type: "note/detail", note: {id: value}}),
+      })}
+      {@const display = switcherFn(type, {
+        r: () => displayRelay({url: value}),
+        p: () => displayPerson(getPersonWithFallback(value)),
+        e: () => "a note",
+        default: "something",
+      })}
+      <div class="mb-4 flex items-center gap-2 border-l-2 border-solid border-gray-5 pl-2">
+        Rated
+        {#if action}
+          <Anchor on:click={action}>{display}</Anchor>
+        {:else}
+          {display}
+        {/if}
+        <div class="text-sm">
+          <Rating inert value={rating} />
+        </div>
+      </div>
+    {/if}
     {#each content as { type, value }, i}
-      {#if type === "newline"}
+      {#if type === "newline" && i > 0}
         {#each value as _}
           <br />
         {/each}
@@ -168,6 +207,11 @@
       {" "}
     {/each}
   </p>
+  {#if invoices.length > 0}
+    <div on:click|stopPropagation>
+      <QRCode fullWidth onClick="copy" code={invoices[0]} />
+    </div>
+  {/if}
   {#if showMedia && links.length > 0}
     <div on:click|stopPropagation>
       <MediaSet {links} />