+++ src/app/views/RelayDetail.svelte
@@ -1,16 +1,32 @@
 <script lang="ts">
-  import {displayRelay, normalizeRelayUrl} from "src/util/nostr"
+  import {batch, timedelta} from "src/util/misc"
+  import {displayRelay, normalizeRelayUrl, getAvgQuality} from "src/util/nostr"
   import Content from "src/partials/Content.svelte"
   import Feed from "src/app/shared/Feed.svelte"
+  import Tabs from "src/partials/Tabs.svelte"
+  import Rating from "src/partials/Rating.svelte"
   import RelayTitle from "src/app/shared/RelayTitle.svelte"
   import RelayActions from "src/app/shared/RelayActions.svelte"
   import {relays} from "src/agent/db"
 
   export let url
 
+  let reviews = []
+  let activeTab = "reviews"
+
   url = normalizeRelayUrl(url)
 
+  $: rating = getAvgQuality("review/relay", reviews)
+
   const relay = relays.get(url) || {url}
+  const tabs = ["reviews", "notes"]
+  const setActiveTab = tab => {
+    activeTab = tab
+  }
+
+  const onReview = batch(1000, chunk => {
+    reviews = reviews.concat(chunk)
+  })
 
   document.title = displayRelay(relay)
 </script>
@@ -20,11 +36,27 @@
     <RelayTitle {relay} />
     <RelayActions {relay} />
   </div>
+  {#if rating}
+    <div class="text-sm">
+      <Rating inert value={rating} />
+    </div>
+  {/if}
   {#if relay.description}
     <p>{relay.description}</p>
   {/if}
-</Content>
-<div class="border-b border-solid border-gray-6" />
-<Content>
-  <Feed relays={[relay]} filter={{kinds: [1]}} />
+  <Tabs borderClass="border-gray-6" {tabs} {activeTab} {setActiveTab} />
+  {#if activeTab === "reviews"}
+    <Feed
+      invertColors
+      onEvent={onReview}
+      delta={timedelta(365, "days")}
+      filter={{
+        kinds: [1985],
+        "#l": ["review/relay"],
+        "#L": ["social.coracle.ontology"],
+        "#r": [relay.url],
+      }} />
+  {:else}
+    <Feed invertColors relays={[relay]} filter={{kinds: [1]}} />
+  {/if}
 </Content>