+++ src/partials/Chip.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
   import cx from "classnames"
 
-  export let theme
+  export let theme = "dark"
+  export let onClick = null
 
   const className = cx($$props.class, "inline-block rounded-full border border-solid py-1 px-2", {
     "border-gray-1": theme === "dark",
@@ -10,8 +11,10 @@
 </script>
 
 <div class={className}>
-  <div class="flex items-center gap-2">
-    <i class="fa fa-times cursor-pointer" on:click|preventDefault />
+  <div class="flex items-center gap-2 whitespace-nowrap">
+    {#if onClick}
+      <i class="fa fa-times cursor-pointer" on:click|preventDefault={onClick} />
+    {/if}
     <slot />
   </div>
 </div>