+++ src/partials/Tabs.svelte
@@ -6,10 +6,11 @@
   export let activeTab
   export let setActiveTab
   export let getDisplay = tab => ({title: toTitle(tab), badge: null})
+  export let borderClass = "border-gray-7"
 </script>
 
 <div
-  class="flex items-center justify-between overflow-auto border-b border-solid border-gray-7 pt-2"
+  class={`flex items-center justify-between overflow-auto border-b border-solid pt-2 ${borderClass}`}
   in:fly={{y: 20}}>
   <div class="flex">
     {#each tabs as tab}
@@ -17,7 +18,7 @@
       <button
         class="flex cursor-pointer gap-2 border-solid border-gray-6 px-8 py-4 hover:border-b"
         class:border-b={activeTab === tab}
-        on:click={() => setActiveTab(tab)}>
+        on:click|preventDefault={() => setActiveTab(tab)}>
         <div>{title}</div>
         {#if badge}
           <div class="h-6 rounded-full bg-gray-6 px-2">{badge}</div>